Added poisoning of Buffer object.
Since buffer objects typically require multiple interactions the
concept of object 'poisoning' was implemented.  If any type of object
transaction fails the object is tagged as poisoned.  This allows the
status of the object to be interpreted at the end of a series of
modifications reducing the number of conditional checks required.