Error capture and poisoning fixes.
The error variable in the internal state structure for the File
object was not being set in all locations where a system call
could fail.  This patch address handles setting of this variable.

A number of the method calls were not properly poisoning the
object on error conditions.  This patch also addresses these
issues.