Clear memory space on Buffer destruction.
The reset method of the Buffer object was clearing the memory up
to the current size of the buffer.  The destructor was not
clearing memory which could have led to a leak of information
from material left in the object.

Modify the ->whack method to clear all of the memory which has
been allocated for the object.