Convert Buffer->print method to use hexdump format.
The previous implementation of the ->print method simply dumped
the contents of the buffer as a continuous stream of hexadecimal
numbers which was useful but not intuitive.

This commit modifies the method to dump the buffer in the common
hexdump format with a sidebar containing ASCII representations of
the buffer contents.