Add method to read a String object from a File.
There was no easy method to read a classis null-terminated C
string from a File object.  This functionality was added in this
change.

The File object is read until it reaches a newline character.
All characters up to this point are placed into the String object
provided by the caller.  The newline character is not returned.