Implement page sized buffering of I/O.
The previous ->slurp and ->read_Buffer methods used decidedly
sub-optimal sizes for the read(2) system calls which were being
issued.  The poorest performance was when a read was done from a
given file position to the end of the file.  In this case single
byte reads were being generated which resulted in poor I/O
performance.

A 4K buffer was defined in the object state structure which is
used as the I/O buffer.  The read_Buffer routine was re-factored
to use 4K reads for all of its I/O.

Since a ->slurp is essentially a read to end of file from the
beginning this method was re-implemented as a thin wrapper around
the ->read_Buffer method.