Define C API linkage for object constructors.
This update provides a basic implementation of support to allow
the HurdLib library to be used with C++ based applications.

The strategy used is to declare the C linkage API with the HCLINK
macro.  This macros is defined to "C" when the header file is
processed by a C++ compiler that has the __cpluscplus define.  In
the absence of such a definition the HCLINK define has a null
definition.