Explicitly define C linkage for Origin object initializer.
This update adds an explicit test for a C++ compiler to the
Origin.h file in order to determine whether or not the
HurdLib_Origin_Init constructor should be defined as having
C linkage.  This is secondary to the fact that code that
implements HurdLib objects needs to include the Origin.h file
which is logically included before the HurdLib.h file.