Implement Gaggle object.
This commit implements an object that manages a collection of
other objects.  The motivation for this object is to have an easy
to use method for managing an array of objects.

Due to type issues the Gaggle.h header file implements three
macros of importance:

GADD(gaggle, object):

Adds an object to a gaggle.

GGET(gaggle, type):

Fetches an object of type from the current cursor position.  Each
call to GGET advances the cursor position one object.  The
->reset method can be used to rewind the cursor to the first
object in the Gaggle

GWHACK(gaggle, type):

Calls the destructor for each object in the Gaggle.  It should be
noted that the WHACK macro needs to be called on the Gaggle
itself to avoid lost object warnings.

The ->size method returns the number of objects in the Gaggle.

In order to safely traverse a list this method should be called
and then the GGET macros should be invoked that many times to
retrieve the objects in the Gaggle.

It is important to note that running off the end of the list
currently results in a null pointer de-reference.  A
'feature' that deserves addressing.