Fix compiler warning in parser.
The compilation of the basic-parser.c file was generating a
warning mesasge secondary to an unused variable.

In order to have a warning free compile, a special target was
added for the compilation of the basic-parser.o object file so
that the -Wno-unused-function command-line flag could be used as
a compiler option.