Change selection of printable characters in Buffer object.
The ->hprint method dumps the contents of a Buffer object in a
classic hexadecimal representation.  The right hand side of the
output was displaying some printable characters as periods rather
then the charcters themselves.

The function used to make a decision as to the nature of the
character was changed from isalpha() to isprint() in order to do
this.