Add method to allow sprintf additions to String object.
This commit adds the following method to the String object:

->add_sprintf

Which allows contents to be added to the String object using the
snprintf function.

The addition is done in a dynamic manner by running the variadic
version of the function in a mode that allows the computation of
the size of the output buffer that is needed.  The underlying
Buffer object that supports the string is increased and the
variadic function is run again pointing to the block of memory
that was added.

This update also adds a unit test for the String object to verify
this functionality.