Implement macro to retrieve an object.
The GGET macro cannot be safely used unless it is known that the
object being retrieved is within the range of the object list.

This commit implements the GPTR macro that is used to retrieve an
object using an opaque pointer retrieved from a Gaggle object.

This macro can be used to detect when a Gaggle traversal has hit
the end of the object list since the pointer returned will be a
NULL value.