Use CO macro to define constant type arguments.
The CO macro is used to define a variable to be constant and what
it points to as constant.

The open declaration of these constant objects was causing newer
compilers to generate warnings about redundant declarations.
Convert all open coding of these arguments to use the CO macro
that silences the error.