Fix compiler error in section name allocation.
Version 9 of the GCC compiler was generating an error message
regarding the use of the strncpy() function that was being used
to copy the section name into the memory that was allocated for
that name.

It is unclear why strncpy() was being used, given the fact that
the incoming name had been validated as a legitimate string and
sufficient memory had been allocated to hold the string.

Convert the use of strncpy to strcpy.