Add File->write_String method.
Implementation work with HurdLib indicated that a common pattern
was the need to write the contents of a String object to a file.

This was being done by copying the contents of the String object
to a Buffer object and then writing the Buffer object with the
File->write_buffer object.

This required a need in some cases to instantiate a Buffer object
and code to conduct the copy of the String object to the Buffer.

To address this the File->write_String method was added localizes
all of this logic in the File based method.