Eradicate the use of the 'auto' keyword.
Due to legacy coding practices the 'auto' keyword had been used
to designate stack based variables.

This practice is not only not useful but collides with a new use
of the keyword auto in recent versions of the C++ standard.

Remove the use of the auto keyword from both the source files and
the code generator.