Update Gaggle implement to be simple array like.
Testing the quixote-export utility indicated that the Gaggle
object was malfunctioning in some way.  In order to address this
the object was reimplemented to make its functionality more like
a classic array than the previous implementation.

On each call to the ->add method a 'slot' is registered in the
array by adding a block of null characters equal in size to a
void pointer.  The address of the object is cast to a void **
pointer that was treated as an array in which the incoming object
should be placed into based on the current size of the array.

The reverse procedure is used on a get.  The address of the
Buffer object that serves as the backing storage for the array is
cast into a void ** pointer that is referenced with the current
value of the cursor.

The Gaggle.h header file was updated so the macros now treat
their arguments as simple pointers.

The GWHACK object was updated to destroy the Gaggle object after
all of the containing objects have been destroyed.  Previous code
that used a separate GWHACK and WHACK can be upgraded at leisure
since the GWHACK macro will set the Gaggle object to NULL causing
it be bypassed by a subsequent WHACK call against the Gaggle
object.