Implement Process object.
In order to simplify the execution a process the Process object
was added.

This object implements the following methods:

Process->add_argument

Process->set_executable

Process->run

Process->run_command

Process->run_command_line

The ->set_executable method must be called first in order to
define the executabl ecode that will be invoked.

The ->add_argument method is used to incrementally add arguments
to the command that is to be invoked.

The ->run method is used to invoke an executable command that has
been setup with a series of arguments.

The ->run_command method is a convenience function used to
automate the execution of the equivalent of a command-line
argument to an executable.

The ->run_command_line method is a convenience function use to
automate the execution of a series of command-line argument that
has the command to be run delimited from the other arguments with
a -- character sequence.

For example, the following command-line argument array:

-X -o option -- executable_name arg1 arg2

When passed to the ->run_command_line method will create and
execute a command of the following form:

executable_name arg1 arg2