Add clear method to File object.
There are instances where the functional status of the File
object needs to be cleared after an I/O error has occurred.
This issue presented itself in the TSEMcontrol object where
writes to the TSEM control file needed to continue after a
control plane write had failed secondary to a process not being
present.

Add the File->clear method which sets the 'poisoned' member of
the object state information to false.