Simplify implementation of write_String method.
The overhead of the File->write_String method was excessively
high given that a Buffer object was being unnecessarily created
in order to write the contents of the String object to the file
descriptor.

Replace the current implementation with an implementation similar
to the ->write_Buffer method which uses the Buffer contents of
the String argument as the source of the I/O, along with the
state size of the String object.