Automate generation of tests and test objects.
The targets to clean the source tree were leaving the object
files associated with the unit tests.

Replace the TESTS variable with a variable name TSRC that
contains the name of the source files for all of the units tests.

The TESTS variable is constructed from the list of test source
files by removal of the .c extension.  In a manner similar to the
generation of the object files generate the object file names for
the units by replacing the .c extension with a .o extension.