Implement macro for initializing object state.
Previously the _init_state() function was implemented in each
HurdLib based object that was responsible for initializing the
state of an object to a set of known values.

These initializations were almost universally setting variables
to NULL values.  This required the need to update the _init_state
function each time a new piece of state information was added, a
process that is error prone.

The OBJ_STATE macro was implemented as a helper macro that
automates the creation of the name of the state structure.

The INIT_STATE macro uses the OBJ_STATE macro to automate the
process of automating the process of initializing the state
structure.  The macros uses memset to zero out the entire
contents of the state structure from the libid member forward.
By convention the libid member is always the first member of a
state structure.